fix the last test and move some files
